# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Delilah Restó API",
    "description": "Esta aplicacion permite interactuar con la iformacion de Delila restó, usuarios, prouctos y pedidos.",
    "contact": {
      "name": "Natalia Camero",
      "url": "https://github.com/nataliacamero/delilah-resto",
      "email": "nataliacameroc@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://www.apache.org/licenses"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "My local server",
      "variables": {
        "port": {
          "default": "3000"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/usuarios": {
      "post": {
        "tags": [
          "Crear, logear y autenticar"
        ],
        "summary": "Crear un nuevo usuario",
        "description": "Permite crear usuarios en la base de datos.",
        "parameters": [
          {
            "name": "usuarios",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Crear un nuevo usuario en la base de datos"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del usuario",
                    "example": "Pablo"
                  },
                  "apellido": {
                    "type": "string",
                    "description": "Apellidos del usuario",
                    "example": "Camero"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email del usuario",
                    "example": "nataliacameroc@gmail.com"
                  },
                  "telefono": {
                    "type": "string",
                    "description": "Telefono del usuario",
                    "example": "3013606833"
                  },
                  "direccionEnvio": {
                    "type": "string",
                    "description": "Direccion de envio del usuario",
                    "example": "Calle 85 # 44 - 77"
                  },
                  "nombreUsuario": {
                    "type": "string",
                    "description": "Nombre de usuario, o correo electronico",
                    "example": "nataliacameroc@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password del usuario",
                    "example": "NAHGhjjygsdf**"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario Creado",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "description": "Nombre del usuario",
                      "example": "Pablo"
                    },
                    "apellido": {
                      "type": "string",
                      "description": "Apellidos del usuario",
                      "example": "Camero"
                    },
                    "email": {
                      "type": "string",
                      "description": "Email del usuario",
                      "example": "nataliacameroc@gmail.com"
                    },
                    "telefono": {
                      "type": "string",
                      "description": "Telefono del usuario",
                      "example": "3013606833"
                    },
                    "direccionEnvio": {
                      "type": "string",
                      "description": "Direccion de envio del usuario",
                      "example": "Calle 85 # 44 - 77"
                    },
                    "nombreUsuario": {
                      "type": "string",
                      "description": "Nombre de usuario, o correo electronico",
                      "example": "nataliacameroc@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "description": "Password del usuario",
                      "example": "NAHGhjjygsdf**"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Peticion sin exito",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 400": "El usuario ya existe en la base de datos."
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuarios existentes",
        "description": "Solo el usuario Administrador lista los usuarios existentes en bd. Debe enviar por headers el token del usuario autorizado.",
        "responses": {
          "200": {
            "description": "Peticion exitosa",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "nombre": {
                        "type": "string",
                        "example": "Natalia"
                      },
                      "apellido": {
                        "type": "string",
                        "example": "Camero"
                      },
                      "email": {
                        "type": "string",
                        "example": "nataliacameroc@gmail.com"
                      },
                      "telefono": {
                        "type": "string",
                        "example": "3013606833"
                      },
                      "direccionEnvio": {
                        "type": "string",
                        "example": "Calle 85 # 44 - 77"
                      },
                      "nombreUsuario": {
                        "type": "string",
                        "description": "Nombre de usuario, o correo electronico",
                        "example": "nataliacameroc@gmail.com"
                      },
                      "password": {
                        "type": "string",
                        "example": "NAHGhjjygsdf**"
                      },
                      "rol": {
                        "type": "string",
                        "example": "Administrador"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2020-05-31T20:40:25.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2020-05-31T20:40:25.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 401. Error al validar el usuario.": null
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Crear, logear y autenticar"
        ],
        "summary": "Autenticar un usuario",
        "description": "Permite a un usuario ya creado autenticarse.",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Autentica un usuario existente en la base de datos."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "nombreUsuario": {
                    "type": "string",
                    "description": "Nombre de usuario, o correo electronico",
                    "example": "nataliacameroc@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password del usuario",
                    "example": "NAHGhjjygsdf**"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario logueado",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwibm9tYnJlIjoiTmF0YWxpYSIsIm5vbWJyZVVzdWFyaW8iOiJuYXRhbGlhY2FtZXJvY0BnbWFpbC5jb20iLCJwYXNzd29yZCI6IktKVUJIWUFTJiYlVFVHWUdZSiIsInJvbCI6IkFkbWluaXN0cmFkb3IiLCJpYXQiOjE1OTA5NTkzMjF9.wum_e9qc4qUtbhvFHEVrnL8WE69sHcMb9dRuA8iGyDM"
              }
            }
          },
          "400": {
            "description": "Peticion sin exito",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 400": "Usuario o contraseña incorrecta."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/seguro": {
      "post": {
        "tags": [
          "Crear, logear y autenticar"
        ],
        "summary": "Autorizar a un usuario",
        "description": "Permite autorizar a un usuario logueado.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Debe poner el token entregado en el metodo login",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwibm9tYnJlIjoiTmF0YWxpYSIsIm5vbWJyZVVzdWFyaW8iOiJuYXRhbGlhY2FtZXJvY0BnbWFpbC5jb20iLCJwYXNzd29yZCI6IktKVUJIWUFTJiYlVFVHWUdZSiIsInJvbCI6IkFkbWluaXN0cmFkb3IiLCJpYXQiOjE1OTA5NTkzMjF9.wum_e9qc4qUtbhvFHEVrnL8WE69sHcMb9dRuA8iGyDM"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario Autenticado",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "200": "Esta es una página autenticada. Hola Natalia. Tu rol es Administrador"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Peticion no exitosa",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 401": "Credeciales invalidas."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Peticion no exitosa",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 401": "Error al validar el usuario."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Lista todos los productos de Delilah Resto",
        "description": "Permite obtener todos los productos.Debe enviarse por headers el token del usuario autorizado.",
        "responses": {
          "200": {
            "description": "Peticion exitosa",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "nombreProducto": {
                        "type": "string",
                        "example": "Hamburguesa Clasica"
                      },
                      "imagen": {
                        "type": "string",
                        "example": "/home/natalita/Escritorio/Proyecto Delilah-Resto/delilah-resto/uploads/134c4cd6-53d8-4b56-9d0e-a2ff0b461f51.png"
                      },
                      "precio": {
                        "type": "number",
                        "example": 6200
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2020-05-31T20:40:25.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2020-05-31T20:40:25.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Peticion sin exito",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 500": "No se puede conectar a la base de datos"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "summary": "Crear productos",
        "description": "Permite crear productos en la base de datos. Debe enviar el token por headers",
        "parameters": [
          {
            "in": "path",
            "name": "productos",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Crear un nuevo producto en bd."
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "id": {
                    "description": "Id del producto",
                    "type": "string",
                    "example": 1
                  },
                  "nombreProducto": {
                    "description": "Nombre del producto",
                    "type": "string",
                    "example": "Hamburguesa Clasica"
                  },
                  "imageFile": {
                    "description": "Imagen del producto",
                    "type": "string",
                    "format": "binary",
                    "example": "imagenes/plato_hamburguesa_clasica"
                  },
                  "precio": {
                    "description": "Precio del producto",
                    "type": "number",
                    "example": 6200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Peticion exitosa",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "description": "Id del producto",
                      "type": "string",
                      "example": 1
                    },
                    "nombreProducto": {
                      "description": "Nombre del producto",
                      "type": "string",
                      "example": "Hamburguesa Clasica"
                    },
                    "imagen": {
                      "description": "Imagen del producto",
                      "type": "string",
                      "format": "binary",
                      "example": "imagenes/plato_hamburguesa_clasica"
                    },
                    "precio": {
                      "description": "Precio del producto",
                      "type": "number",
                      "example": 6200
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2020-05-31T20:40:25.000Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2020-05-31T20:40:25.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Peticion sin exito",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "Error 500": "No se puede conectar a la base de datos."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}